% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{dendrogram_server}
\alias{dendrogram_server}
\title{Dendrogram Server}
\usage{
dendrogram_server(input, output, session, height = "1000px", resetButton)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{height}{height of dendrogram}

\item{resetButton}{resetButton (actionbutton)}
}
\value{
shiny server
}
\description{
Dendrogram Server
}
