% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTree.R
\name{makeTree}
\alias{makeTree}
\title{Fast version of collapsibleTree.Node from collapsibleTree package}
\usage{
makeTree(df, hierarchy_attribute = "level", root = df$name,
  inputId = NULL, attribute = "leafCount", aggFun = sum,
  fill = "lightsteelblue", linkLength = NULL, fontSize = 10,
  tooltip = FALSE, tooltipHtml = NULL, nodeSize = NULL,
  collapsed = TRUE, zoomable = TRUE, width = NULL, height = NULL,
  ...)
}
\arguments{
\item{df}{see https://github.com/AdeelK93/collapsibleTree}

\item{hierarchy_attribute}{see https://github.com/AdeelK93/collapsibleTree}

\item{root}{see https://github.com/AdeelK93/collapsibleTree}

\item{inputId}{see https://github.com/AdeelK93/collapsibleTree}

\item{attribute}{see https://github.com/AdeelK93/collapsibleTree}

\item{aggFun}{see https://github.com/AdeelK93/collapsibleTree}

\item{fill}{see https://github.com/AdeelK93/collapsibleTree}

\item{linkLength}{see https://github.com/AdeelK93/collapsibleTree}

\item{fontSize}{see https://github.com/AdeelK93/collapsibleTree}

\item{tooltip}{see https://github.com/AdeelK93/collapsibleTree}

\item{tooltipHtml}{see https://github.com/AdeelK93/collapsibleTree}

\item{nodeSize}{see https://github.com/AdeelK93/collapsibleTree}

\item{collapsed}{see https://github.com/AdeelK93/collapsibleTree}

\item{zoomable}{see https://github.com/AdeelK93/collapsibleTree}

\item{width}{see https://github.com/AdeelK93/collapsibleTree}

\item{height}{see https://github.com/AdeelK93/collapsibleTree}

\item{...}{see https://github.com/AdeelK93/collapsibleTree}
}
\value{
see https://github.com/AdeelK93/collapsibleTree
}
\description{
Fast version of collapsibleTree.Node from collapsibleTree package
}
